% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intro_slides.R
\name{intro_slides}
\alias{intro_slides}
\title{Intro Slides}
\usage{
intro_slides(...)
}
\arguments{
\item{...}{Arguments passed to ioslides_presentation.}
}
\description{
Creates introduction ioslides markdown template.
}
