% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meetup_import.R
\name{meetup_import}
\alias{meetup_import}
\title{Import Meetup .csv}
\usage{
meetup_import(file = NULL, ...)
}
\arguments{
\item{file}{File path to .csv file. Defaults to NULL. Default attempts to import 
any .csv file present in the \strong{data-raw/meetup} folder of the working directory.
See \strong{Details} section and file argument of \code{\link{read.csv}}}

\item{...}{arguments passed to \code{\link{read.csv}}}
}
\value{
A \code{\link{data.frame}} formated indentically to \code{\link{member_list}}.
}
\description{
Import, Clean, and Format of data manually downloaded from meetup.com's 
member list page
}
\details{
Assumes manual download of the meetup member list file. For example, the 
Chicago R User Group (CRUG) url looks like this: https://www.meetup.com/ChicagoRUG/members/?op=csv
Note: One can only download this file if signed into meetup.com as an organizer.

The file download from meetup.com is in .xls format. To ensure proper conversion 
to .csv, open it with excel and then `save as .csv`. Alternatively, one can 
upload the .xls file to google sheets and then export as .csv.
  
To utilize the Default file path of this function, place the file the folder 
path \strong{`data-raw/meetup`} of the working directory and make sure it is
the only file in the folder.
}
\examples{
\dontrun{ 
member_list <- meetup_import()}
}
\seealso{
\code{\link{read.csv}}
}
