% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_mem_counter.R
\name{new_mem_counter}
\alias{new_mem_counter}
\title{New Member Counter}
\usage{
new_mem_counter(data)
}
\arguments{
\item{data}{A \code{\link{data.frame}} created by exporting the members list from meetup.com.
Note: All date fields must by coverted to class \code{\link{Date}}.}
}
\value{
\code{new_mem_counter} returns a \code{\link{data.frame}} containing 2 variables. 
These are the `Date` of each meetup, and how many `New` members joined leading up to each 
meetup.
}
\description{
Generates totals of new members between each meetup
}
\examples{
\dontrun{
member_file_path <- paste0(getwd(),"vignettes/data/ChicagoRUG_Member_List_on_04-09-18.csv")

member_list <- read.csv(member_file_path, stringsAsFactors=FALSE)

 # Convert date columns to type 'Date'.
 member_list$Joined.Group.on <- as.Date.character(member_list$Joined.Group.on, format = "\%m/\%d/\%Y")
 member_list$Last.Attended <- as.Date.character(member_list$Last.Attended, format = "\%m/\%d/\%Y")

 new_members <- new_mem_counter(member_list)
 }
 
}
