% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pizza_estimate.R
\name{pizza_estimate}
\alias{pizza_estimate}
\title{Estimate pizza order}
\usage{
pizza_estimate(registered = 120, pizza_diameter = 18, attend_rate = 0.6,
  serving = 2, style = "thincrust")
}
\arguments{
\item{registered}{Integer. The number of people registered on Meetup.com}

\item{pizza_diameter}{Integer. Diameter of the largest pizza available in inches. 
Defaults to 18 inches for an extra large pizza.}

\item{attend_rate}{Numeric of type double. The percentage estimate of attendance. 
Defaults to 60 percent attendance rate, expressed as \strong{0.60}.}

\item{serving}{Integer or Numeric. Number of servings per person. 
Default is 2 and does not have to be a whole number. See references for methodology.}

\item{style}{Character vector of either "thincrust", "pan", or "deepdish". 
Defaults to Chicago "thincrust".}
}
\value{
\code{pizza_estimate} returns a \code{\link{data.frame}} containing 5 variables. 
These are \code{registered, est_attend, eaters_per_pizza, style, and the pizza_estimate}.
}
\description{
Function to compute how many pizzas to order based on arguments provided. 
Defaults to \strong{thin-crust} style.
}
\examples{
pizza_estimate(registered=150, pizza_diameter=18, attend_rate=0.60, serving = 2, 
style = "thincrust")

}
\references{
Lisa R. Young, Ph.D., R.D. \href{https://www.nestleusa.com/enjoypizza}{Nestle USA Site: Enjoy Pizza}
and \href{https://www.nestleusa.com/asset-library/documents/nutritionhealthwellness/pizza/pizzaportionguide_full.pdf}{PDF: The Fun of Pizza, the Balance of Good Nutrition, page 5.}
}
