% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pizza_quantity.R
\name{pizza_quantity}
\alias{pizza_quantity}
\title{Estimate pizza quantity to order}
\usage{
pizza_quantity(registered, pizza_diameter = 18, attend_rate = 0.55,
  pizza_eat = 49)
}
\arguments{
\item{registered}{Integer of the number of people registered on Meetup.com}

\item{pizza_diameter}{Interger representing diameter in inches of the largest 
pizza available. Defaults to 18 or 18 inces for an extra large}

\item{attend_rate}{Numeric of type double representing a percentage estimate 
of attendance. Defaults to 55 percent attendance, input 0.60.}

\item{pizza_eat}{Integer representing square inches a member consumers.Defaults to 49 or 7 by 7 inches.}
}
\value{
\code{pizza_quantity} returns a \code{\link{data.frame}} containing 4 variables. 
These are \code{registered, estimated_attendance, people_per_pizza, pizza_estimate}.
}
\description{
How many pizzas should we order?
}
\examples{

pizza_quantity(130, pizza_diameter=18, attend_rate=0.65)

}
