% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_list.R
\docType{data}
\name{member_list}
\alias{member_list}
\title{member_list}
\format{A data.frame with 3101 observations on 17 variables:
\itemize{
 \item \strong{Member.ID}
 \item \strong{Joined.Group.on}
 \item \strong{Last.visited.group.on}
 \item \strong{Last.Attended}
 \item \strong{Total.RSVPs}
 \item \strong{RSVPed.Yes}
 \item \strong{RSVPed.Maybe}
 \item \strong{RSVPed.No}
 \item \strong{Meetups.attended}
 \item \strong{No.shows}
 \item \strong{Intro}
 \item \strong{Photo}
 \item \strong{Assistant.Organizer}
 \item \strong{Mailing.List}
 \item \strong{R.Experience}
 \item \strong{Count.Index}
 }}
\source{
\url{https://www.meetup.com/ChicagoRUG/members/?op=csv}
}
\usage{
data(member_list)
}
\description{
Chicago R User Group Data downloaded from meetup.com/ChicagoRUG. PID removed, data formated.
}
\examples{
 str(member_list)
}
\keyword{datasets}
