% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_mem_counter.R
\name{new_mem_counter}
\alias{new_mem_counter}
\title{New Member Counter}
\usage{
new_mem_counter(data)
}
\arguments{
\item{data}{A \code{\link{data.frame}} created by exporting the members list from meetup.com.
Note: All date fields must by coverted to class \code{\link{Date}}.}
}
\value{
\code{new_mem_counter} returns a \code{\link{data.frame}} containing 2 variables. 
These are the `Date` of each meetup, and how many `New` members joined leading up to each 
meetup.
}
\description{
Generates sum of new members between each meetup.
}
\examples{
 data(member_list)

 new_members <- new_mem_counter(member_list)
 
 tail(new_members)
 
 # Plot the `new members` data since 2017.
 
 # Create Date Range Index
 Range <- new_members[new_members$Date > "2017-01-01" & new_members$Date <= Sys.Date(),]
 Date_Index <- as.numeric(row.names(Range))
 
 # Create x-axis labels, using year-month date format
 x_labels <- format(new_members$Date[Date_Index], "\%Y-\%m")
 
 # Plot
 barplot(new_members$New[Date_Index], names.arg = x_labels, las=2,
         main = "CRUG members, joined between meetups",  
         ylab = "New Members", xlab = "")
         
}
