% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pizza_quantity.R
\name{pizza_quantity}
\alias{pizza_quantity}
\title{Estimate pizza order}
\usage{
pizza_quantity(registered = 120, pizza_diameter = 18, attend_rate = 0.6,
  serving = 2)
}
\arguments{
\item{registered}{Integer. The number of people registered on Meetup.com}

\item{pizza_diameter}{Integer. Diameter of the largest pizza available in inches. 
Defaults to 18 inches for an extra large pizza.}

\item{attend_rate}{Numeric of type double. The percentage estimate of attendance. 
Defaults to 60 percent attendance rate, expressed as \strong{0.60}.}

\item{serving}{Integer or Numeric. Number of servings per person. 
Default is 2 and does not have to be a whole number. See references for methodology.}
}
\value{
\code{pizza_quantity} returns a \code{\link{data.frame}} containing 4 variables. 
These are \code{registered, estimated_attendance, people_per_pizza, and the pizza_estimate}.
}
\description{
Function to compute how many \strong{thin-crust} pizzas to order based on arguments provided.
}
\details{
\strong{WARNING: }This function is for thin-crust pizza only! Do not use for deep-dish or pan-pizza.
}
\examples{
pizza_quantity(registered=150, pizza_diameter=18, attend_rate=0.65, serving = 2)

}
\references{
Lisa R. Young, Ph.D., R.D. \href{https://www.nestleusa.com/enjoypizza}{Nestle USA Site: Enjoy Pizza}
and \href{https://www.nestleusa.com/asset-library/documents/nutritionhealthwellness/pizza/pizzaportionguide_full.pdf}{PDF: The Fun of Pizza, the Balance of Good Nutrition, page 5.}
}
